Conflicts:
	automation/Build.py
	automation/Common.py
	automation/framework/utils.py
	src/CMakeLists.txt
	src/backend/barrier/CMakeLists.txt
	src/backend/passes/BackendPassesWrappers/CMakeLists.txt
	src/backend/tests/opencl/Reference/cl/work_group_broadcast.tst
	src/backend/tests/opencl/Reference/cl/work_group_broadcast2d.tst
	src/backend/tests/opencl/Reference/cl/work_group_broadcast3d.tst
	src/backend/utils/CompilationUtils.cpp
	src/backend/validations/ImathLibd/CMakeLists.txt
	src/backend/validations/NEAT/CMakeLists.txt
	src/backend/validations/OCLKernelDataGenerator/CMakeLists.txt
	src/backend/validations/REFALU/CMakeLists.txt
	src/fe_compilers/clang_compiler/CMakeLists.txt
	src/fe_compilers/clang_compiler/clang_compiler.cpp
	src/fe_compilers/clang_compiler/clang_driver.cpp
	src/fe_compilers/clang_compiler/clang_driver.h
	src/tests/conform1.2/src/build_conform.py
	src/tests/conform1.2/src/test_common/harness/imageHelpers.cpp

+++ opencl-intel/framework/d3d9_sharing/d3d9_resource.cpp
@@ -239,7 +239,7 @@ size_t D3D9Surface::GetMemObjSize() const
     const D3DSURFACE_DESC desc = GetDesc(*GetResourceInfo());
     const UINT plane = dynamic_cast<const D3D9SurfaceResourceInfo*>(GetResourceInfo())->m_plane;
     const cl_image_format clFormat = MapD3DFormat2OclFormat(desc.Format, plane);
-    switch (desc.Format)
+    switch ((unsigned long int)desc.Format)
     {
     case MAKEFOURCC('N', 'V', '1', '2'):
         assert(0 == plane || 1 == plane);
